Devember 2 - Output
Today we handled output.  We also cleaned up a lot of code by moving it
out into directories to keep things organized.

Possible output options are defined by the configuration, but the
context of the request (namely the Accept header) helps decide which of
the outputs to choose from.  By default this is JSON, and we add other
outputs as needed.

The outputs should handle different data types easily and in ways that
make sense for the output type.