Devember 4 - Cleanup and UI
Spend a bit of time doing some preliminary cleanup of how Commands are
registered, though we are still doing it manually.  We should probably
tackle this at some point.

We also decided to build out a basic UI so we can get that started. We
still want to build the route creator.  So we need to be able to
dynamically add new routes and have the system handle that
appropriately.

Yep, that's about it.