Devember 5 - Configuration and Route display
So, didn't really think about it before, but decided to add
configuration injection into our commands.  This seems to have worked
out well and I'm happy with the results.  The command defines the
configuration and where it expects it, and the application provides the
external configuration in the appropriate location and it automatically
gets injected.

We also now can display route details.  This still needs more work, but
we are doing well and progressing.

Yay!