Devember 11 - Router Builder Start
More cleanup today, but more importantly, started working on the Router
Builder.  This is the magic part of it all.  The part that makes it
possible to create routers and merge together commands without actually
having to write any code.  So it's getting close to be even more fun.

Happy to say as well that the build stuff I built up yesterday on
Devember 10 worked on my Windows machine without issue.  So I can
successfully run this on both my Windows and Mac without issue.  So
that's a plus.

Note: spending an hour working on this might not seem like much, but
what I'm able to accomplish is still encouraging.